tests/backend/challenge/modules/task/services/RetrieveTaskByIdServiceTest.java
package backend.challenge.modules.task.services;


import backend.challenge.modules.task.dtos.TaskDTO;
import backend.challenge.modules.task.models.Task;
import backend.challenge.modules.task.repositories.ITaskRepository;
import backend.challenge.modules.task.repositories.TaskRepository;
import kikaha.core.test.KikahaRunner;
import org.junit.Before;
import org.junit.Test;
import org.junit.runner.RunWith;

import static org.junit.Assert.assertEquals;
import static org.junit.Assert.assertTrue;

@RunWith( KikahaRunner.class )
public class RetrieveTaskByIdServiceTest {

	private IRetrieveTaskByIdService retrieveTaskByIdService;
	private UtilTestsTasks utilTestsTasks;
	private CreateTaskService createTaskService;

	@Before
	public void init() {
		final ITaskRepository taskRepository = new TaskRepository();
		utilTestsTasks = new UtilTestsTasks();
		retrieveTaskByIdService = new RetrieveTaskByIdService(taskRepository);
		createTaskService = new CreateTaskService(taskRepository);
	}

	@Test
	public void shouldBeAbleToListTheTaskById() {
		TaskDTO taskDTO = utilTestsTasks.createTask();
		Task task = createTaskService.execute(taskDTO);
		Task testTask = retrieveTaskByIdService.execute(task.getId());

		assertEquals(testTask.getId(), task.getId());
	}

}
